+++ src/Plugin.Settings.Net45/Settings.cs
@@ -11,7 +11,7 @@ namespace Plugin.Settings
     /// </summary>
     public class SettingsImplementation : ISettings
     {
-        static IsolatedStorageFile Store => IsolatedStorageFile.GetUserStoreForAssembly();
+        static IsolatedStorageFile Store => IsolatedStorageFile.GetUserStoreForDomain();
         
 
         private readonly object locker = new object();
@@ -421,5 +421,11 @@ namespace Plugin.Settings
             AddOrUpdateValueInternal(key, value, fileName);
 
         #endregion
+
+        /// <summary>
+        /// Attempts to open the app settings page.
+        /// </summary>
+        /// <returns>true if success, else false and not supported</returns>
+        public bool OpenAppSettings() => false;
     }
 }