+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 3.1.1
+* Add ability to open app settings page in iOS/Android
+* Android: Better parsing of double's with exponents. #101
+
 ### 3.0.0
 * Strongly Type all Gets and Sets
 * Re-targeted to .NET Standard 1.0